+++ R/mirt.R
@@ -460,9 +460,7 @@
 #'   the default number of quasi-Monte Carlo integration nodes will be set to 5000 in total
 #' @param TOL convergence threshold for EM or MH-RM; defaults are .0001 and .001. If
 #'   \code{SE.type = 'SEM'} and this value is not specified, the default is set to \code{1e-5}.
-#'   If \code{dentype = 'empiricalhist'} (i.e., \code{'EH'}) or \code{'empiricalhist_Woods'} (i.e., \code{'EHW'})
-#'   and \code{TOL} is not specified then the default \code{3e-5}
-#'   will be used. To evaluate the model using only the starting values pass \code{TOL = NaN}, and
+#'   To evaluate the model using only the starting values pass \code{TOL = NaN}, and
 #'   to evaluate the starting values without the log-likelihood pass \code{TOL = NA}
 #' @param dentype type of density form to use for the latent trait parameters. Current options include
 #'